% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_ps_termid}
\alias{calc_ps_termid}
\title{Calculate 4 digit PowerSchool TermID given first year of a school year}
\usage{
calc_ps_termid(sy, quarter = 0)
}
\arguments{
\item{sy}{First year in a school year (e.g. 2015 for SY 2015-2016)}

\item{quarter}{quarter the quarter as integer in a year. The default is 0,
which returns the school year's termid.}
}
\value{
a character or integer vector
}
\description{
Calculate 4 digit PowerSchool TermID given first year of a school year
}
\examples{
\dontrun{calc_ps_termid(2015)}
}
