% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{check_get_connection}
\alias{check_get_connection}
\title{Check if DB connections available and valid or create new one}
\usage{
check_get_connection(.database_name)
}
\arguments{
\item{.database_name}{name of the database you want to connect to}
}
\value{
returns an S4 object that inherits from DBIConnection.
This object is used to communicate with the database engine.

Note that this function is called for it's side-effect: it will create
a connection object with the name \code{conn_\{.database_name\}} so that other
functions and users have access to the connection.
}
\description{
Check if DB connections available and valid or create new one
}
