% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinit.R
\name{check_macos_linux}
\alias{check_macos_linux}
\title{Check if OS is *Nix flavored}
\usage{
check_macos_linux()
}
\value{
a single entry boolean (i.e., `TRUE` or `FALSE`)
}
\description{
Check if OS is *Nix flavored
}
