% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{create_connection}
\alias{create_connection}
\title{Create connection to database}
\usage{
create_connection(.database_name, .server_name, env = globalenv())
}
\arguments{
\item{.database_name}{name of the database you want to connect to}

\item{.server_name}{the name of the server the database is hosted on}

\item{env}{which environment to save the connection.  The default is the global
environment and you should not change this unless you really need to and you know
what you are doing.}
}
\value{
returns an S4 object that inherits from DBIConnection.
This object is used to communicate with the database engine.

Note that this function is called for it's side-effect: it will create
a connection object with the name \code{conn_\{.database_name\}} so that other
functions and users have access to the connection.
}
\description{
Create connection to database
}
\examples{
\dontrun{create_connection("PROD1")}
}
