% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_locations}
\alias{db_locations}
\title{Locations of databases in IDEA's data warehouse}
\format{
A data frame with 132 rows and 3 variables:
\describe{
\item{server_name}{the server name that hosts the data base in \code{database_name}}
\item{database_name}{the name of a database, natch}
\item{url}{the IP Address of the server for the data base in \code{database_name}}
...
}
}
\usage{
db_locations
}
\description{
A dataset containing the server name, IP address of all IDEA databases
}
\keyword{datasets}
