% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{generate_schema}
\alias{generate_schema}
\title{Generates server.database.dbo schema string}
\usage{
generate_schema(.database_name)
}
\arguments{
\item{.database_name}{Name of the database your want to connect to}
}
\value{
a string
}
\description{
Generates server.database.dbo schema string
}
