% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AEES.R
\name{get_aees}
\alias{get_aees}
\alias{get_aees_oe_questions}
\alias{get_aees_oe_responses}
\alias{get_aees_questions}
\alias{get_aees_results}
\alias{get_aees_sections}
\title{Connect to \code{ResultsView} View or other tables  in \code{AEES} on \code{REDACTED-SQLSERVER} which provides
a tidy (i.e. a long ) view of non-OE responses from the 2020-21 AEES survey}
\usage{
get_aees()

get_aees_oe_questions()

get_aees_oe_responses()

get_aees_questions()

get_aees_results()

get_aees_sections()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{ResultsView} View or other tables  in \code{AEES} on \code{REDACTED-SQLSERVER} which provides
a tidy (i.e. a long ) view of non-OE responses from the 2020-21 AEES survey
}
\section{Functions}{
\itemize{
\item \code{get_aees_oe_questions()}: \code{OEQuestions} table in \code{AEES} on \code{REDACTED-SQLSERVER}

\item \code{get_aees_oe_responses()}: \code{OEResponses} table in \code{AEES} on \code{REDACTED-SQLSERVER}

\item \code{get_aees_questions()}: \code{Questions} table in \code{AEES} on \code{REDACTED-SQLSERVER}

\item \code{get_aees_results()}: \code{Results} table in \code{AEES} on \code{REDACTED-SQLSERVER}

\item \code{get_aees_sections()}: \code{Sections} table in \code{AEES} on \code{REDACTED-SQLSERVER}

}}
\examples{
# This attaches to the school db with \code{conn_AEES} connection
\dontrun{
aees <- get_aees()
}

}
