% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD1.R
\name{get_currently_enrolled_students}
\alias{get_currently_enrolled_students}
\title{Connect to `Students` table on \code{PROD1} in `Schools` schema only pull
currently enrolled students}
\usage{
get_currently_enrolled_students()
}
\value{
a `tbl_sql SQL Server` object.
}
\description{
Specifically, pull students data where `EnrollmentStatus == 0`,
`RowIsCurrent = TRUE` and `SchoolTermID = max(SchoolTermID)`
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection
current <- get_currently_enrolled_students()

}
