% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GPTW.R
\name{get_gptw}
\alias{get_gptw}
\alias{get_gptw_oe_questions}
\alias{get_gptw_oe_responses}
\alias{get_gptw_questions}
\alias{get_gptw_results}
\alias{get_gptw_sections}
\title{Connect to \code{ResultsView} View or other tables  in \code{GPTW} on \code{REDACTED-SQLSERVER} which provides
a tidy (i.e. a long ) view of non-OE responses from the 2020-21 GPTW survey}
\usage{
get_gptw()

get_gptw_oe_questions()

get_gptw_oe_responses()

get_gptw_questions()

get_gptw_results()

get_gptw_sections()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{ResultsView} View or other tables  in \code{GPTW} on \code{REDACTED-SQLSERVER} which provides
a tidy (i.e. a long ) view of non-OE responses from the 2020-21 GPTW survey
}
\section{Functions}{
\itemize{
\item \code{get_gptw_oe_questions()}: \code{OEQuestions} table in \code{GPTW} on \code{REDACTED-SQLSERVER}

\item \code{get_gptw_oe_responses()}: \code{OEResponses} table in \code{GPTW} on \code{REDACTED-SQLSERVER}

\item \code{get_gptw_questions()}: \code{Questions} table in \code{GPTW} on \code{REDACTED-SQLSERVER}

\item \code{get_gptw_results()}: \code{Results} table in \code{GPTW} on \code{REDACTED-SQLSERVER}

\item \code{get_gptw_sections()}: \code{Sections} table in \code{GPTW} on \code{REDACTED-SQLSERVER}

}}
\examples{
# This attaches to the school db with \code{conn_GPTW} connection
\dontrun{
gptw <- get_gptw()
}

}
