% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD2.R
\name{get_iabwa}
\alias{get_iabwa}
\title{Connect to `IABWA` table on \code{PROD2} in `Assessments` schema}
\usage{
get_iabwa()
}
\value{
a `tbl_sql SQL Server` object.
}
\description{
Connect to `IABWA` table on \code{PROD2} in `Assessments` schema
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection
schools <- get_schools()

# This pulls down schools data form the DB and onto your computer
# and then cleans the names (lower snakecase) using [janitor::clean_names()]
\dontrun{
ias_bwas <- get_ia_bwa() \%>\%
  filter(AcademicYear == "2020-2021") \%>\%
  collect() \%>\%
  clean_names()
  }
}
