% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LearnLocation.R
\name{get_learning_location}
\alias{get_learning_location}
\title{Connect to \code{LearningLocation} database \code{791150-hqvra} which provides
daily learning location (i.e, remote or in-person) for every student on ever day}
\usage{
get_learning_location()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{LearningLocation} database \code{791150-hqvra} which provides
daily learning location (i.e, remote or in-person) for every student on ever day
}
\details{
The learning location of a student  on a given is determined by calculating a
30-day average  ("days" includes only days attended) in each location (remote vs in-person).
75\% of those 30 days at home leads to an \code{att_category} of remote, 75\% of those
30 days at home leads to an \code{att_category} of at-home, and all other cases are
coded mixed.
}
\examples{
# This attaches to the school db with \code{conn_Dashboard} connection
gptw <- get_gptw()
}
