% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Persistence.R
\name{get_persistence_historical}
\alias{get_persistence_historical}
\title{Connect to \code{PersistenceHistorical} table in \code{Persistence} database on \code{REDACTED-SQLSERVER}}
\usage{
get_persistence_historical()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{PersistenceHistorical} table in \code{Persistence} database on \code{REDACTED-SQLSERVER}
}
\examples{
# This attaches to the `Persistence` db with \code{conn_Persistence} connection
\dontrun{
ps_hist <- get_persistence_historical()
}
}
