% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Persistence.R
\name{get_persistence_reasons_codes}
\alias{get_persistence_reasons_codes}
\title{Connect to \code{PersistenceReasonsCodes}  table in \code{Persistence} database on \code{RGVPDSD-DWPRDS}}
\usage{
get_persistence_reasons_codes()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{PersistenceReasonsCodes}  table in \code{Persistence} database on \code{RGVPDSD-DWPRDS}
}
\examples{
# This attaches to the `Persistence` db with \code{conn_Persistence} connection
\dontrun{
ps_reason_codes <- PersistenceReasonsCodes()
}
}
