% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_servers.R
\name{get_renstar_math}
\alias{get_renstar_math}
\alias{get_renstar_reading}
\title{Connect to \code{StarMathV2} or \code{StarReadingV2} tables in \code{SRC_AR} on \code{964592-SQLDS}}
\usage{
get_renstar_math()

get_renstar_reading()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{StarMathV2} or \code{StarReadingV2} tables in \code{SRC_AR} on \code{964592-SQLDS}
}
\details{
Not that this is a linked server connection from  \code{791150-HQVRA}
(i.e., R&A's SQL Server instance) to \code{964592-SQLDS}.  Since we need to connect
to a database when making a connection we connect to the \code{Dashboard} database
on \code{964592-SQLDS}.  Doing so results in the conn_Dashboard connection object,
which is fine, but a bit of a hack.
}
\section{Functions}{
\itemize{
\item \code{get_renstar_reading}: Connect to \code{StarReadingV2} table in \code{SRC_AR} on \code{964592-SQLDS}
}}

\examples{
# This attaches to the school db with \code{conn_Dashboard} connection
\dontrun{
rs_math <- get_renstar_math()
}
}
