% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SchoolsExtensions.R
\name{get_schools_extensions}
\alias{get_schools_extensions}
\title{Connect to \code{dbo.SchoolsExtensions} View or other tables  in \code{lk_Schools} on \code{RGVPDSD-DWPRD2} which provides
a tidy (i.e. a long ) view of schools information}
\usage{
get_schools_extensions()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{dbo.SchoolsExtensions} View or other tables  in \code{lk_Schools} on \code{RGVPDSD-DWPRD2} which provides
a tidy (i.e. a long ) view of schools information
}
\examples{
# This attaches to the school db with \code{conn_lk_Schools} connection
\dontrun{
schools_extensions <- get_schools_extensions()
}

}
