% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD1.R
\name{get_student_daily_attendance}
\alias{get_student_daily_attendance}
\title{Connect to `Students` table on \code{PROD1} in `Attendance` schema which has
membership (i.e., enrollment) and attendance data}
\usage{
get_student_daily_attendance()
}
\value{

}
\description{
Connect to `Students` table on \code{PROD1} in `Attendance` schema which has
membership (i.e., enrollment) and attendance data
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection

# This pulls down schools data form the DB and onto your computer
# and then cleans the names (lower snakecase) using [janitor::clean_names()]
stu_attend <- get_student_daily_attendance()

}
