% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD1.R
\name{get_students}
\alias{get_students}
\title{Connect to `Students` table on \code{PROD1} in `Schools` schema}
\usage{
get_students()
}
\value{
a `tbl_sql SQL Server` object.
}
\description{
Connect to `Students` table on \code{PROD1} in `Schools` schema
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection
stus <- get_students()

# This pulls down schools data form the DB and onto your computer
# and then cleans the names (lower snakecase) using [janitor::clean_names()]

\dontrun{
schools <- get_schools() \%>\%
  collect() \%>\%
  clean_names()
  }
}
