% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse_meta_data.R
\name{get_warehouse_meta_data}
\alias{get_warehouse_meta_data}
\title{Gets warehouse meta data from R&A Documentation data base}
\usage{
get_warehouse_meta_data()
}
\value{
a tibble, which is written to \code{namespace:ideadata}
}
\description{
Gets warehouse meta data from R&A Documentation data base
}
