% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{id_tables_in_dbs}
\alias{id_tables_in_dbs}
\title{Identify rows with given table name, database name, or schema in the MetaData table in the R&A's documentation
database.}
\usage{
id_tables_in_dbs(.table_name, .database_name = NULL, .schema = NULL)
}
\arguments{
\item{.table_name}{the name of a table that may be located somewhere, as a quoted string}

\item{.database_name}{the name of a database in the warehouse , as a quoted string}

\item{.schema}{the name of a schema in the warehouse, as a quoted string}
}
\value{
a tibble with all unique table, schema, database, and server combinations found in the MetaData table
}
\description{
Identify rows with given table name, database name, or schema in the MetaData table in the R&A's documentation
database.
}
\examples{
\dontrun{
library(ideadata)

id_tables_in_dbs("Students")
}
}
