% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinit.R
\name{kinit}
\alias{kinit}
\title{Checks/Gets Kereberos Ticket-Getting-Tickets are}
\usage{
kinit(uid, pwd)
}
\arguments{
\item{uid}{your IDEA user id (just the `firstname.lastname` part)}

\item{pwd}{your IDEA pass word}
}
\value{
None, this function is run for its side-effects
}
\description{
Checks/Gets Kereberos Ticket-Getting-Tickets are
}
