% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_creds}
\alias{setup_creds}
\title{Interactively update .Renviron file}
\usage{
setup_creds()
}
\value{
the path to the .Renviron file, invisibly.  This function is called for
its side-effects
}
\description{
Adds entries to your `.Renviron` file for uid, password, and driver.
}
\details{
Running `setup_creds()` will check for for your .Revniron file, create it if
is missing, and add then environment files necessary for [ideadata] to function.
Not that if (if the file is empty, it will delete it and create a new `.Renviron`
file).
}
\examples{

setup_creds()
}
