% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_uid}
\alias{update_uid}
\alias{update_pwd}
\alias{update_driver}
\title{Update IDEA_RNA_DB_UID in .Renviron file}
\usage{
update_uid(.path)

update_pwd(.path)

update_driver(.path)
}
\arguments{
\item{.path}{path to .Renviron file}
}
\value{
Called for side-effect of updating .Renviron
}
\description{
Update IDEA_RNA_DB_UID in .Renviron file
}
\section{Functions}{
\itemize{
\item \code{update_pwd()}: Update IDEA_RNA_DB_PWD in .Renviron file

\item \code{update_driver()}: Update IDEA_RNA_ODBC_DRIVER in .Renviron file

}}
