% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_warehouse_metada}
\alias{view_warehouse_metada}
\title{View servers, schemas, databases, and tables currently available in IDEA's data warehouse}
\usage{
view_warehouse_metada()
}
\value{
Nothing, this function is called for it's side effects (name's opeing a sortable/filterable Viewer)
}
\description{
View servers, schemas, databases, and tables currently available in IDEA's data warehouse
}
