% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{warehouse_meta_data}
\alias{warehouse_meta_data}
\title{table level details of all databases in IDEA's data warehouse}
\format{
A data frame and 4 variables:
\describe{
\item{server_name}{the server name that hosts the data base in \code{database_name}}
\item{database_name}{the name of a database, natch}
\item{schema}{the name of a schema, natch}
\item{table_name}{the name of a table, natch}
}
}
\source{
generated from MetaData view in the R&A SQL Server instance
}
\description{
table level details of all databases in IDEA's data warehouse
}
\details{
A dataset containing the fields and their locations (tables, databases, servers)
}
