% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector.R
\name{collector}
\alias{collector}
\title{Collect data from database to local environment piecemeal}
\usage{
collector(.data, .split_column)
}
\arguments{
\item{.data}{remote \code{\link[dbplyr:tbl]{dbplyr::tbl()}} table that needs to be collected}

\item{.split_column}{column used to break \code{.data} into peices to be downloaded}
}
\value{
a tibble
}
\description{
Collect data from database to local environment piecemeal
}
\details{
IDEA has 60,000+ students, which means data that is collected daily
or more frequently can get large really quickly as the day goes along.
}
\examples{
library(dplyr)

regions_remote <- get_regions()
regions <- regions_remote \%>\% collector(State)

}
