% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect (close) a connection from IDEA's data warehouse}
\usage{
disconnect(connection_name)
}
\arguments{
\item{con}{the name  \link[DBI:DBIConnection-class]{DBI::DBIConnection} object \strong{as a string}}
}
\value{
nothing, as it's called for its side-effects
}
\description{
Disconnect (close) a connection from IDEA's data warehouse
}
\details{
This is a thing wrapper around \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}, which closes
the connection, discards all pending work, and frees resources (e.g., memory, sockets).
}
\examples{
# The following creats a connect call conn_PROD1 in global environment
regions <- get_regions()

#notice the name is a character string and not the bare object name
disconnect("conn_PROD1")
}
