% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from a database in the IDEA data warehouse}
\usage{
disconnect(con)
}
\arguments{
\item{con}{the name  \link[DBI:DBIConnection-class]{DBI::DBIConnection} object \strong{as a string}}
}
\value{
nothing, as it's called for its side-effects
}
\description{
Disconnect from a database in the IDEA data warehouse
}
\details{
This is a thing wrapper around \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}, which closes
the connection, discards all pending work, and frees resources (e.g., memory, sockets).
}
\examples{
# The following creates a connect call `conn_PROD1` in global environment
\dontrun{
regions <- get_regions()

disconnect(conn_PROD1)
}
}
