% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD1.R
\name{get_currently_enrolled_students}
\alias{get_currently_enrolled_students}
\title{Connect to \code{Students} table on \code{PROD1} in \code{Schools} schema only pull
currently enrolled students}
\usage{
get_currently_enrolled_students()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Specifically, pull students data where \code{EnrollmentStatus == 0},
\code{RowIsCurrent = TRUE} and \code{SchoolTermID = max(SchoolTermID)}
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection
current <- get_currently_enrolled_students()

}
