% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_db_url}
\alias{get_db_url}
\title{retrieves database URL and server name}
\usage{
get_db_url(.database_name)
}
\arguments{
\item{.database_name}{}
}
\value{
a data frame, with on road providing the `server_name`, `database_name`
and `url` of the database given by `.database_name`
}
\description{
retrieves database URL and server name
}
\examples{
get_db_url("PROD1")
}
