% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Persistence.R
\name{get_persistence_code_historical}
\alias{get_persistence_code_historical}
\title{Connect to \code{PersitenceCodeHistorical} (\emph{sic}) table in \code{Persistence} database on \code{RGVPDSD-DWPRD1}}
\usage{
get_persistence_code_historical()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{PersitenceCodeHistorical} (\emph{sic}) table in \code{Persistence} database on \code{RGVPDSD-DWPRD1}
}
\examples{
# This attaches to the `Persistence` db with \code{conn_Persistence} connection
\dontrun{
ps_code <- get_persistence_code()
}
}
