% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_servers.R
\name{get_persistence_code_historical}
\alias{get_persistence_code_historical}
\title{Connect to \code{PersitenceCodeHistorical} (\emph{sic}) table in \code{Persistence} database on \code{964592-SQLDS}}
\usage{
get_persistence_code_historical()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{PersitenceCodeHistorical} (\emph{sic}) table in \code{Persistence} database on \code{964592-SQLDS}
}
\details{
Not that this is a linked server connection from  \code{791150-HQVRA}
(i.e., R&A's SQL Server instance) to \code{964592-SQLDS}.  Since we need to connect
to a linked database when making a connection to this table we first connect to the \code{Dashboard} database
on \code{964592-SQLDS}.  Doing so results in the conn_Dashboard connection object,
which is fine, but a bit of a hack.
}
\examples{
# This attaches to the `Persistence` db with \code{conn_Dashboard} connection
\dontrun{
ps_code <- get_persistence_code()
}
}
