% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD1.R
\name{get_regions}
\alias{get_regions}
\title{Connect to `Regions` table on \code{PROD1} in `Schools` schema}
\usage{
get_regions()
}
\value{

}
\description{
Connect to `Regions` table on \code{PROD1} in `Schools` schema
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection
stus <- get_students()

# This pulls down schools data form the DB and onto your computer
# and then cleans the names (lower snakecase) using [janitor::clean_names()]

\dontrun{
regions <- get_regions() \%>\%
  collect() \%>\%
  clean_names()
  }
}
