% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RenStar.R
\name{get_renstar_math}
\alias{get_renstar_math}
\alias{get_renstar_reading}
\title{Connect to \code{StarMathV2} or \code{StarReadingV2} tables in \code{SRC_AR} on \code{RGVPDSD-DWSRC4}}
\usage{
get_renstar_math()

get_renstar_reading()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{StarMathV2} or \code{StarReadingV2} tables in \code{SRC_AR} on \code{RGVPDSD-DWSRC4}
}
\section{Functions}{
\itemize{
\item \code{get_renstar_reading}: Connect to \code{StarReadingV2} table in \code{SRC_AR} on \code{RGVPDSD-DWSRC4}
}}

\examples{
# This attaches to the school db with \code{conn_SRC_AR} connection
\dontrun{
rs_math <- get_renstar_math()
}
}
