% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PROD1.R
\name{get_schools}
\alias{get_schools}
\title{Connect to \code{Schools} table on \code{PROD1} in \code{Schools} schema}
\usage{
get_schools()
}
\value{
a \verb{tbl_sql SQL Server} object.
}
\description{
Connect to \code{Schools} table on \code{PROD1} in \code{Schools} schema
}
\examples{

# This attaches to the school db with \code{conn_PROD1} connection
schools <- get_schools()

# This pulls down schools data form the DB and onto your computer
# and then cleans the names (lower snakecase) using [janitor::clean_names()]
\dontrun{
library(dplyr)
library(janitor)
schools <- get_schools() \%>\%
  collect() \%>\%
  clean_names()
  }
}
