% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_creds}
\alias{setup_creds}
\title{Interactively update .Renviron file}
\usage{
setup_creds()
}
\value{
the path to the .Renviron file, invisibly.  This function is called for
its side-effects
}
\description{
Adds entries to your \code{.Renviron} file for uid, password, and driver.
}
\details{
Running \code{setup_creds()} will check for  your .Revniron file, create it if
is missing, and add the environment files necessary for {ideadata} to function.
Not that if the \code{.Renviron} file is empty, the script will delete it and create a new \code{.Renviron}
file before populating with the uid, pwd, and driver variables.
}
\examples{

\dontrun{setup_creds()}
}
