% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warehouse_meta_data}
\alias{warehouse_meta_data}
\title{Field level details of all (R&A and non-R&A databases) in IDEA's data warehouse}
\format{
A data frame with 69,591 rows and 7 variables:
\describe{
\item{server_name}{the server name that hosts the data base in \code{database_name}}
\item{database_name}{the name of a database, natch}
\item{table_name}{the name of a table, natch}
\item{column_name}{the name of a column, natch}
\item{data_type}{the column's data_type in SQL Server (not R)}
\item{is_nullable}{can the column contain NULL values?}
\item{url}{the IP Address of the server for the data base in \code{database_name}}
...
}
}
\usage{
warehouse_meta_data
}
\description{
A dataset containing the fields and their locations (tables, databases, servers)
}
\keyword{datasets}
